+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'src/blocs/movies_bloc.dart';
-import 'src/ui/movies_list.dart';
+import 'src/ui/home_page.dart';
 
 void main() {
   runApp(App());
@@ -12,7 +12,7 @@ class App extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: MoviesList(
+      home: HomePage(
           title: "Movies App",
           bloc: _moviesBloc,
         ),
