+++ lib/src/blocs/movies_bloc.dart
@@ -1,26 +1,26 @@
 import 'dart:async';
-import '../models/item_model.dart';
+import '../models/movie_model.dart';
 import '../resources/movies_repository.dart';
 import 'i_movies_bloc.dart';
 
 class MoviesBloc extends IMoviesBloc {
-  MoviesRepository moviesRepository = MoviesRepository();
-  StreamController<ItemModel> moviesStreamController = StreamController();
+  MoviesRepository _moviesRepository = MoviesRepository();
+  StreamController<MovieModel> _moviesStreamController = StreamController();
 
   @override
   void initialize() {}
 
   @override
   void dispose() {
-    this.moviesStreamController.close();
+    _moviesStreamController.close();
   }
 
   @override
-  Stream<ItemModel> get moviesStream => this.moviesStreamController.stream;
+  Stream<MovieModel> get moviesStream => _moviesStreamController.stream;
 
   @override
   void fetchAllMovies() async {
-    final _moviesList = await moviesRepository.fetchAllMovies();
-    this.moviesStreamController.sink.add(_moviesList,);
+    final _moviesList = await _moviesRepository.fetchAllMovies();
+    _moviesStreamController.sink.add(_moviesList);
   }
 }
