+++ lib/src/models/movie_result.dart
@@ -1,31 +1,4 @@
-class ItemModel {
-  int _page;
-  int _total_results;
-  int _total_pages;
-  List<_Result> _results = [];
-
-  ItemModel.fromJson(Map<String, dynamic> json) {
-    _page = json['page'];
-    _total_results = json['total_results'];
-    _total_pages = json['total_pages'];
-    List<_Result> temp = [];
-    for (int i = 0; i < json['results'].length; i++) {
-      _Result result = _Result(json['results'][i]);
-      temp.add(result);
-    }
-    _results = temp;
-  }
-
-  List<_Result> get results => _results;
-
-  int get total_pages => _total_pages;
-
-  int get total_results => _total_results;
-
-  int get page => _page;
-}
-
-class _Result {
+class MovieResult {
   int _vote_count;
   int _id;
   bool _video;
@@ -41,7 +14,7 @@ class _Result {
   String _overview;
   String _release_date;
 
-  _Result(result) {
+  MovieResult(result) {
     _vote_count = result['vote_count'];
     _id = result['id'];
     _video = result['video'];
