+++ lib/src/resources/movies_repository.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 import 'movies_api_provider.dart';
-import '../models/item_model.dart';
+import '../models/movie_model.dart';
 
 class MoviesRepository {
   final moviesApiProvider = MoviesApiProvider();
 
-  Future<ItemModel> fetchAllMovies() => moviesApiProvider.fetchMoviesList();
+  Future<MovieModel> fetchAllMovies() => moviesApiProvider.fetchMoviesList();
 }
