+++ lib/utils/constants.dart
@@ -7,5 +7,5 @@ abstract class Constants {
   static const String uriClient =
       "https://api.themoviedb.org/3/movie/popular?api_key=385737c6d715026a1cd89526ff9d8d44";
   static const int status = 200;
-  static const String error = 'I failed the load';
+  static const String error = 'An error has ocurred while loading movies';
 }
